User-Service/requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Supabase
supabase>=2.0.0
gotrue>=2.0.0

# Validation and serialization
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# JWT handling
python-jose[cryptography]>=3.3.0
PyJWT>=2.8.0

# Environment and config
python-dotenv>=1.0.0

# Logging
structlog>=24.1.0

# HTTP client
httpx>=0.26.0

# Phone number validation
phonenumbers>=8.13.0
